+++ ACP Monitor.lua
@@ -9,36 +9,23 @@
 -- Resolve Repository Root
 --------------------------------------------------------------------------------
 
-local source =
-    debug.getinfo(1, "S").source
-
-local scriptPath =
-    source:sub(2):match("(.*/)")
-
-local repositoryRoot =
-    scriptPath:gsub("/$", "")
+local source = debug.getinfo(1, "S").source
+local scriptPath = source:sub(2):match("(.*/)")
+local repositoryRoot = scriptPath:gsub("/$", "")
 
 --------------------------------------------------------------------------------
 -- Configure Package Path
 --------------------------------------------------------------------------------
 
-local modulePaths =
-{
+local modulePaths = {
     repositoryRoot .. "/?.lua",
     repositoryRoot .. "/?/init.lua"
 }
 
 for _, modulePath in ipairs(modulePaths) do
-
     if not package.path:find(modulePath, 1, true) then
-
-        package.path =
-            package.path
-            .. ";"
-            .. modulePath
-
+        package.path = package.path .. ";" .. modulePath
     end
-
 end
 
 --------------------------------------------------------------------------------
@@ -48,9 +35,9 @@ end
 local Bootstrap =
     require("Core.Application.Bootstrap")
 
-Bootstrap.Initialize(
-    repositoryRoot
-)
+assert(
+    Bootstrap.Initialize(),
+    "Unable to initialize ACP Monitor.")
 
 --------------------------------------------------------------------------------
 -- Application
@@ -60,5 +47,4 @@ local MonitorApplication =
     require("Core.Application.Monitor.MonitorApplication")
 
 MonitorApplication.Initialize()
-
 MonitorApplication.Run()